feat: add support for Auto IAM AuthN
AlloyDB uses a metadata exchange prior to handing over to the database
driver to begin the Postgres protocol. As part of this metadata
exchange, the dialer sends the IAM principal's OAuth2 token which the
proxy server uses to verify connection permissions and, if IAM AuthN is
requested, to authenticate to the database.